# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with Gin, GORM, and Swagger — v1 uses custom query params, v2 uses OData",
        "title": "Go API V6",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:3020",
    "basePath": "/",
    "paths": {
        "/api/v1/products": {
            "get": {
                "description": "Returns a paginated list of products with optional sorting, filtering, and search",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to return (default: 10)",
                        "name": "top",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default: 0)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field name to sort by (e.g. name, price)",
                        "name": "sortby",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction: asc or desc (default: asc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term matched against product name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by field value (e.g. filter[name]=Laptop)",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Repository.PaginationResult-Models_Product"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new product with name and price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a product",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "description": "Returns a single product by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated product data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a product by ID (soft delete via GORM)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "description": "Returns a paginated list of users with optional sorting, filtering, and search",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to return (default: 10)",
                        "name": "top",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default: 0)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field name to sort by (e.g. name, email)",
                        "name": "sortby",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction: asc or desc (default: asc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term matched against name and email",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by field value (e.g. filter[name]=Herris)",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Repository.PaginationResult-Models_User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new user with name and email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "description": "Returns a single user by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated user data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a user by ID (soft delete via GORM)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v2/products": {
            "get": {
                "description": "Returns a paginated product collection using OData v4 query options",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get all products (OData v4)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to return (default: 10)",
                        "name": "$top",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default: 0)",
                        "name": "$skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field and direction (e.g. name asc, price desc)",
                        "name": "$orderby",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "OData filter (e.g. name eq 'Laptop' and price gt 10000, contains(name,'lap'), startswith(name,'L'))",
                        "name": "$filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term matched against product name",
                        "name": "$search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ODataPagedResponse-Models_Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/products/$count": {
            "get": {
                "description": "Returns the total number of products as plain text",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product count (OData v4)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OData filter expression",
                        "name": "$filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/products/{id}": {
            "get": {
                "description": "Returns a single product with OData v4 context",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product by ID (OData v4)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/users": {
            "get": {
                "description": "Returns a paginated user collection using OData v4 query options",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users (OData v4)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to return (default: 10)",
                        "name": "$top",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default: 0)",
                        "name": "$skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field and direction (e.g. name asc, email desc)",
                        "name": "$orderby",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "OData filter (e.g. name eq 'Herris', contains(email,'@example.com'))",
                        "name": "$filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term matched against name and email",
                        "name": "$search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ODataPagedResponse-Models_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/users/$count": {
            "get": {
                "description": "Returns the total number of users as plain text",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user count (OData v4)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OData filter expression",
                        "name": "$filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/tutorial_go_Handler_V2.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Models.Product": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                },
                "Price": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "Repository.PaginationResult-Models_Product": {
            "type": "object",
            "properties": {
                "Data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Models.Product"
                    }
                },
                "Skip": {
                    "type": "integer"
                },
                "Top": {
                    "type": "integer"
                },
                "Total": {
                    "type": "integer"
                },
                "TotalPages": {
                    "type": "integer"
                }
            }
        },
...
        "tutorial_go_Handler_V1.CreateProductRequest": {
            "type": "object",
            "required": [
                "Name",
                "Price"
            ],
            "properties": {
                "Name": {
                    "type": "string",
                    "example": "Laptop"
                },
                "Price": {
                    "type": "number",
                    "example": 15000000
                }
            }
        },
...
        "tutorial_go_Handler_V1.ErrorResponse": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "tutorial_go_Handler_V1.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string",
                    "example": "Laptop Pro"
                },
                "Price": {
                    "type": "number",
                    "example": 18000000
                }
            }
        },
...
    }
}